+++ crates/path_resolver/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.4](https://github.com/conda/rattler/compare/path_resolver-v0.2.3...path_resolver-v0.2.4) - 2025-12-18
+
+### Other
+
+- update README.md with new banner image ([#1926](https://github.com/conda/rattler/pull/1926))
+
+## [0.2.3](https://github.com/conda/rattler/compare/path_resolver-v0.2.2...path_resolver-v0.2.3) - 2025-12-08
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.2.2](https://github.com/conda/rattler/compare/path_resolver-v0.2.1...path_resolver-v0.2.2) - 2025-10-28
 
 ### Other