+++ crates/rattler_cache/CHANGELOG.md
@@ -49,6 +49,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.6](https://github.com/conda/rattler/compare/rattler_cache-v0.6.5...rattler_cache-v0.6.6) - 2025-12-18
+
+### Other
+
+- update README.md with new banner image ([#1926](https://github.com/conda/rattler/pull/1926))
+
+## [0.6.5](https://github.com/conda/rattler/compare/rattler_cache-v0.6.4...rattler_cache-v0.6.5) - 2025-12-08
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.6.4](https://github.com/conda/rattler/compare/rattler_cache-v0.6.3...rattler_cache-v0.6.4) - 2025-11-27
 
 ### Other