+++ crates/rattler_config/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.23](https://github.com/conda/rattler/compare/rattler_config-v0.2.22...rattler_config-v0.2.23) - 2025-12-18
+
+### Other
+
+- update README.md with new banner image ([#1926](https://github.com/conda/rattler/pull/1926))
+
+## [0.2.22](https://github.com/conda/rattler/compare/rattler_config-v0.2.21...rattler_config-v0.2.22) - 2025-12-08
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.2.21](https://github.com/conda/rattler/compare/rattler_config-v0.2.20...rattler_config-v0.2.21) - 2025-11-27
 
 ### Other