+++ crates/rattler_lock/CHANGELOG.md
@@ -44,6 +44,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.9](https://github.com/conda/rattler/compare/rattler_lock-v0.26.8...rattler_lock-v0.26.9) - 2025-12-18
+
+### Fixed
+
+- always emit depends in index.json ([#1932](https://github.com/conda/rattler/pull/1932))
+
+### Other
+
+- update README.md with new banner image ([#1926](https://github.com/conda/rattler/pull/1926))
+
+## [0.26.8](https://github.com/conda/rattler/compare/rattler_lock-v0.26.7...rattler_lock-v0.26.8) - 2025-12-08
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.26.7](https://github.com/conda/rattler/compare/rattler_lock-v0.26.6...rattler_lock-v0.26.7) - 2025-11-27
 
 ### Added