+++ crates/rattler_menuinst/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.42](https://github.com/conda/rattler/compare/rattler_menuinst-v0.2.41...rattler_menuinst-v0.2.42) - 2025-12-18
+
+### Other
+
+- update README.md with new banner image ([#1926](https://github.com/conda/rattler/pull/1926))
+
+## [0.2.41](https://github.com/conda/rattler/compare/rattler_menuinst-v0.2.40...rattler_menuinst-v0.2.41) - 2025-12-08
+
+### Added
+
+- Add support for `info_plist_extra` property in menuinst ([#1911](https://github.com/conda/rattler/pull/1911))
+
 ## [0.2.40](https://github.com/conda/rattler/compare/rattler_menuinst-v0.2.39...rattler_menuinst-v0.2.40) - 2025-11-27
 
 ### Fixed